requirements.txt
torch
opencv-python-headless
numpy
pillow
pandas
matplotlib
tqdm
pyyaml
scipy
pycocotools